+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - Added media option to parallax component
   - Added activeClass option to tooltip component
   - Updated tooltip component to look for an updated title attribute before show
+  - Updated timepicker to be initialized on a parent element
   - Updated slideset filter activation
   - Updated changed event from focus.uk.slider to focusitem.uk.slider
   - Removed warp mode from sortable component